feat: 함수의 기능 분리
Person.order 함수 분리 및 일부 수정
