+++ CodeStarterCamp_Week3/Coffee.swift
@@ -1,0 +1,15 @@
+//
+//  Coffee.swift
+//  CodeStarterCamp_Week3
+//
+//  Created by Dean on 2023/03/17.
+//
+
+import Foundation
+
+enum Coffee: String {
+    case americano = "아메리카노"
+    case caffeMocha = "카페모카"
+    case latte = "라떼"
+    case caramelMacchiato = "카라멜 마키아또"
+}
