+++ CodeStarterCamp_Week3/CoffeeShop.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-enum Coffee: String {
-    case americano = "아메리카노"
-    case caffeMocha = "카페모카"
-    case latte = "라떼"
-    case caramelMacchiato = "카라멜 마키아또"
-}
-
 class CoffeeShop {
     var sales: Int = 0
     var menu: [Coffee: Int]
@@ -29,9 +22,17 @@ class CoffeeShop {
         self.barista = barista
     }
     
+    func sell(_ coffee: Coffee) {
+        if let coffeePrice = self.menu[coffee] {
+            sales += coffeePrice
+        }
+    }
+    
     func make(_ coffee: Coffee, for name: String) {
+        print("결제가 완료됐습니다")
+        sell(coffee)
         print("커피를 추출하고 있습니다")
         pickUpTable.append(coffee)
-        print("\(name) 님이 주문하신 \(coffee)(이/가) 준비되었습니다. 픽업대에서 가져가주세요.")
+        print("\(name) 님이 주문하신 \(coffee.rawValue)(이/가) 준비되었습니다. 픽업대에서 가져가주세요.")
     }
 }
