+++ CodeStarterCamp_Week3/Person.swift
@@ -11,18 +11,23 @@ struct Person {
     let name: String
     var money: Int
     
-    mutating func order(_ coffee: Coffee, of coffeeShop: CoffeeShop, by name: String) {
+    func checkPrice(_ coffee: Coffee, of coffeeShop: CoffeeShop) -> Int {
         if let coffeePrice = coffeeShop.menu[coffee] {
-            if money < coffeePrice {
-                print("잔액이 \(coffeePrice - money)원만큼 부족합니다.")
-            } else {
-                money -= coffeePrice
-                coffeeShop.sales += coffeePrice
-                print("결제 완료됐습니다.")
-                coffeeShop.make(coffee, for: name)
-            }
+            return coffeePrice
         } else {
-            print("존재하지 않는 메뉴입니다")
+            print("존재하지 않는 메뉴입니다.")
+            return 0
+        }
+    }
+    
+    mutating func order(_ coffee: Coffee, of coffeeShop: CoffeeShop, by name: String) {
+        let coffeePrice = checkPrice(coffee, of: coffeeShop)
+        
+        if money < coffeePrice {
+            print("잔액이 \(coffeePrice - money)원만큼 부족합니다.")
+        } else {
+            money -= coffeePrice
+            coffeeShop.make(coffee, for: name)
         }
     }
 }
