+++ CodeStarterCamp_Week3/main.swift
@@ -11,6 +11,6 @@ import Foundation
 var misterLee = Person(name: "Lee", money: 20000)
 var missKim = Person(name: "Kim", money: 30000)
 
-var yagombucks = CoffeeShop(menu: [Coffee.americano: 2500, Coffee.caramelMacchiato: 3500, Coffee.latte: 3000, Coffee.caffeMocha: 3000], barista: misterLee)
+var yagombucks = CoffeeShop(menu: [.americano: 2500, .caramelMacchiato: 3500, .latte: 3000, .caffeMocha: 3000], barista: misterLee)
 
-missKim.order(Coffee.americano, of: yagombucks, by: "missKim")
+missKim.order(.americano, of: yagombucks, by: missKim.name)
